+++ default_status.js
@@ -350,19 +350,23 @@ module.exports = [{
   "is_saved": "true"
 }, {
   "id": "2CBDAE96-EC22-48B4-A369-BFC624463C5F",
-  "type": "track","created_by_name": "Sisyphus Industries",
+  "type": "track",
+  "created_by_name": "Sisyphus Industries",
   "name": "Erase",
   "is_deletable": "false",
   "default_vel": 3,
   "default_accel": 0.5,
-  "default_thvmax": 1
+  "default_thvmax": 1,
+  "reversible": "true"
 }, {
   "id": "C3D8BC17-E2E1-4D6D-A91F-80FBB65620B8",
-  "type": "track","created_by_name": "Sisyphus Industries",
+  "type": "track",
+  "created_by_name": "Sisyphus Industries",
   "name": "Tensig 1",
   "default_vel": 1,
   "default_accel": 0.5,
-  "default_thvmax": 1
+  "default_thvmax": 1,
+  "reversible": "true"
 }, {
   "id": "2B34822B-0A27-4398-AE19-23A3C83F1220",
   "type": "track",
@@ -378,7 +382,8 @@ module.exports = [{
   "name": "Circam 2S",
   "default_vel": 1,
   "default_accel": 0.5,
-  "default_thvmax": 1
+  "default_thvmax": 1,
+  "reversible": "true"
 }, {
   "id": "B7407A2F-04C3-4C92-B907-4C3869DA86D6",
   "type": "track",
@@ -386,7 +391,8 @@ module.exports = [{
   "name": "C Warp 3B",
   "default_vel": 1,
   "default_accel": 0.5,
-  "default_thvmax": 1
+  "default_thvmax": 1,
+  "reversible": "true"
 }, {
   "id": "7C046710-9F19-4423-B291-7394996F0913",
   "type": "track",
@@ -402,7 +408,8 @@ module.exports = [{
   "name": "Hep",
   "default_vel": 1,
   "default_accel": 0.5,
-  "default_thvmax": 1
+  "default_thvmax": 1,
+  "reversible": "true"
 }, {
   "id": "26FBFB10-4BC7-46BF-8D55-85AA52C19ADF",
   "type": "track",
@@ -418,40 +425,45 @@ module.exports = [{
   "name": "Para 2B",
   "default_vel": 1,
   "default_accel": 0.5,
-  "default_thvmax": 1
+  "default_thvmax": 1,
+  "reversible": "true"
 }, {
-  "id": "906de7a3-4769-4964-96cd-2bcf80bad2f5",
-  "type": "track",
-  "created_by_name": "Sisyphus Industries",
-  "name": "apache1",
-  "default_vel": 1,
-  "default_accel": 0.5,
-  "default_thvmax": 1
-}, {
+//   "id": "906de7a3-4769-4964-96cd-2bcf80bad2f5",
+//   "type": "track",
+//   "created_by_name": "Sisyphus Industries",
+//   "name": "apache1",
+//   "default_vel": 1,
+//   "default_accel": 0.5,
+//   "default_thvmax": 1,
+//   "reversible": "true"
+// }, {
   "id": "b19a2909-a9f5-490b-9f9a-d990f5a77393",
   "type": "track",
   "created_by_name": "Sisyphus Industries",
   "name": "e10",
   "default_vel": 1,
   "default_accel": 0.5,
-  "default_thvmax": 1
+  "default_thvmax": 1,
+  "reversible": "true"
 }, {
-  "id": "6796b5c5-0f4b-47e2-b871-f30d9e967d6b",
-  "type": "track",
-  "created_by_name": "Sisyphus Industries",
-  "name": "ellweave1$p",
-  "default_vel": 1,
-  "default_accel": 0.5,
-  "default_thvmax": 1
-}, {
-  "id": "bd89c7f7-d272-44dd-b80c-aa4c1ece36f2",
-  "type": "track",
-  "created_by_name": "Sisyphus Industries",
-  "name": "fox1warp1",
-  "default_vel": 1,
-  "default_accel": 0.5,
-  "default_thvmax": 1
-}, {
+//   "id": "6796b5c5-0f4b-47e2-b871-f30d9e967d6b",
+//   "type": "track",
+//   "created_by_name": "Sisyphus Industries",
+//   "name": "ellweave1$p",
+//   "default_vel": 1,
+//   "default_accel": 0.5,
+//   "default_thvmax": 1,
+//   "reversible": "true"
+// }, {
+//   "id": "bd89c7f7-d272-44dd-b80c-aa4c1ece36f2",
+//   "type": "track",
+//   "created_by_name": "Sisyphus Industries",
+//   "name": "fox1warp1",
+//   "default_vel": 1,
+//   "default_accel": 0.5,
+//   "default_thvmax": 1,
+//   "reversible": "true"
+// }, {
   "id": "c0f63e2f-e63d-4228-a8bc-6a29f92da50c",
   "type": "track",
   "created_by_name": "Sisyphus Industries",
@@ -466,7 +478,8 @@ module.exports = [{
   "name": "lasso1",
   "default_vel": 1,
   "default_accel": 0.5,
-  "default_thvmax": 1
+  "default_thvmax": 1,
+  "reversible": "true"
 }, {
   "id": "db96d725-09b9-4c80-80ec-9c77658a74ac",
   "type": "track",
@@ -474,23 +487,26 @@ module.exports = [{
   "name": "lasso4a",
   "default_vel": 1,
   "default_accel": 0.5,
-  "default_thvmax": 1
+  "default_thvmax": 1,
+  "reversible": "true"
 }, {
-  "id": "45540bac-3422-4386-a804-2b185776d921",
-  "type": "track",
-  "created_by_name": "Sisyphus Industries",
-  "name": "peoni1",
-  "default_vel": 1,
-  "default_accel": 0.5,
-  "default_thvmax": 1
-}, {
+//   "id": "45540bac-3422-4386-a804-2b185776d921",
+//   "type": "track",
+//   "created_by_name": "Sisyphus Industries",
+//   "name": "peoni1",
+//   "default_vel": 1,
+//   "default_accel": 0.5,
+//   "default_thvmax": 1,
+//   "reversible": "true"
+// }, {
   "id": "f1451810-ab23-4878-9b45-cca5cfb5b1e7",
   "type": "track",
   "created_by_name": "Sisyphus Industries",
   "name": "seash5",
   "default_vel": 1,
   "default_accel": 0.5,
-  "default_thvmax": 1
+  "default_thvmax": 1,
+  "reversible": "true"
 }, {
   "id": "580c9f06-1c74-4e36-b4ab-b63a72d6399b",
   "type": "track",
@@ -500,21 +516,23 @@ module.exports = [{
   "default_accel": 0.5,
   "default_thvmax": 1
 }, {
-  "id": "3872a018-f176-4661-b580-a988e0788299",
-  "type": "track",
-  "created_by_name": "Sisyphus Industries",
-  "name": "sola3",
-  "default_vel": 1,
-  "default_accel": 0.5,
-  "default_thvmax": 1
-}, {
+//   "id": "3872a018-f176-4661-b580-a988e0788299",
+//   "type": "track",
+//   "created_by_name": "Sisyphus Industries",
+//   "name": "sola3",
+//   "default_vel": 1,
+//   "default_accel": 0.5,
+//   "default_thvmax": 1,
+//   "reversible": "true"
+// }, {
   "id": "103f52ff-fc00-4c41-9098-47f2e0a7e7a1",
   "type": "track",
   "created_by_name": "Sisyphus Industries",
   "name": "spmrph1r",
   "default_vel": 1,
   "default_accel": 0.5,
-  "default_thvmax": 1
+  "default_thvmax": 1,
+  "reversible": "true"
 }, {
   "id": "1f274aa7-6214-4172-b251-a5ac33d36184",
   "type": "track",
@@ -522,7 +540,8 @@ module.exports = [{
   "name": "swirl1b",
   "default_vel": 1,
   "default_accel": 0.5,
-  "default_thvmax": 1
+  "default_thvmax": 1,
+  "reversible": "true"
 }, {
   "id": "e0a61484-35eb-4f40-9a0a-7f9d6ee58c37",
   "type": "track",
@@ -530,7 +549,8 @@ module.exports = [{
   "name": "tensig2$",
   "default_vel": 1,
   "default_accel": 0.5,
-  "default_thvmax": 1
+  "default_thvmax": 1,
+  "reversible": "true"
 }, {
   "id": "3ae084de-cea3-4729-8bf8-0c6e43d47d49",
   "type": "track",
@@ -538,7 +558,8 @@ module.exports = [{
   "name": "trialdn1",
   "default_vel": 1,
   "default_accel": 0.5,
-  "default_thvmax": 1
+  "default_thvmax": 1,
+  "reversible": "true"
 }, {
   "id": "6209824a-9860-4746-a374-d02091e9f808",
   "type": "track",
@@ -546,7 +567,8 @@ module.exports = [{
   "name": "trwirl1",
   "default_vel": 1,
   "default_accel": 0.5,
-  "default_thvmax": 1
+  "default_thvmax": 1,
+  "reversible": "true"
 }, {
   "id": "77a9f396-f2a3-459b-8e0d-5da024eb13b6",
   "type": "track",
@@ -554,12 +576,13 @@ module.exports = [{
   "name": "zowie1r$b",
   "default_vel": 1,
   "default_accel": 0.5,
-  "default_thvmax": 1
+  "default_thvmax": 1,
+  "reversible": "true"
 }, {
   "id": "cfca6c0c-0d04-4fbc-a78b-3d031fd6434f",
   "type": "track",
   "created_by_name": "Sisyphus Industries",
-  "name": "Sailboat2 ",
+  "name": "Sailboat2",
   "default_vel": 1,
   "default_accel": 0.5,
   "default_thvmax": 1
@@ -776,7 +799,8 @@ module.exports = [{
   "name": "fox1warp2",
   "default_vel": 1,
   "default_accel": 0.5,
-  "default_thvmax": 1
+  "default_thvmax": 1,
+  "reversible": "true"
 }, {
   "id": "6a27135c-e93b-4925-9dd9-d138b42719f0",
   "type": "track",
@@ -784,7 +808,8 @@ module.exports = [{
   "name": "apache1",
   "default_vel": 1,
   "default_accel": 0.5,
-  "default_thvmax": 1
+  "default_thvmax": 1,
+  "reversible": "true"
 }, {
   "id": "dcaeaee1-446a-4880-a942-32ef05edcd2a",
   "type": "track",
@@ -792,7 +817,8 @@ module.exports = [{
   "name": "ellweave1$p",
   "default_vel": 1,
   "default_accel": 0.5,
-  "default_thvmax": 1
+  "default_thvmax": 1,
+  "reversible": "true"
 }, {
   "id": "4e718106-4005-4623-825c-9e3da30a2109",
   "type": "track",
@@ -800,7 +826,8 @@ module.exports = [{
   "name": "peoni1",
   "default_vel": 1,
   "default_accel": 0.5,
-  "default_thvmax": 1
+  "default_thvmax": 1,
+  "reversible": "true"
 }, {
   "id": "38cb2dad-23d6-4c53-888f-88ee591a04d6",
   "type": "track",
@@ -808,7 +835,8 @@ module.exports = [{
   "name": "sola3",
   "default_vel": 1,
   "default_accel": 0.5,
-  "default_thvmax": 1
+  "default_thvmax": 1,
+  "reversible": "true"
 }, {
   "id": "452e628f-fe10-4858-9ed0-fc4b47b6d2ff",
   "type": "playlist",